+++ jhpt-server/src/main/java/com/tyj/jhpt/server/body/dto/SuperDto.java
@@ -12,43 +12,43 @@ package com.tyj.jhpt.server.body.dto;
  */
 public class SuperDto {
     // 最高电压电池系统号
-    private byte maxSystemNo;
+    private int maxSystemNo;
     // 最高电压电池单体代号
-    private byte maxSimpleNo;
+    private int maxSimpleNo;
     // 电池单体电压最高值
     private int max;
     // 最低电压电池系统号
-    private byte minSystemNo;
+    private int minSystemNo;
     // 最低电压电池单体代号
-    private byte minSimpleNo;
+    private int minSimpleNo;
     // 电池单体电压最低值
     private int min;
     // 最高温度子系统号
-    private byte maxTemperatureSystemNo;
+    private int maxTemperatureSystemNo;
     // 最高温度探针序号
-    private byte maxSeq;
+    private int maxSeq;
     // 最高温度值
-    private byte maxTemperature;
+    private int maxTemperature;
     // 最低温度子系统号
-    private byte minTemperatureSystemNo;
+    private int minTemperatureSystemNo;
     // 最低温度探针序号
-    private byte minSeq;
+    private int minSeq;
     // 最低温度值
-    private byte minTemperature;
+    private int minTemperature;
 
-    public byte getMaxSystemNo() {
+    public int getMaxSystemNo() {
         return maxSystemNo;
     }
 
-    public void setMaxSystemNo(byte maxSystemNo) {
+    public void setMaxSystemNo(int maxSystemNo) {
         this.maxSystemNo = maxSystemNo;
     }
 
-    public byte getMaxSimpleNo() {
+    public int getMaxSimpleNo() {
         return maxSimpleNo;
     }
 
-    public void setMaxSimpleNo(byte maxSimpleNo) {
+    public void setMaxSimpleNo(int maxSimpleNo) {
         this.maxSimpleNo = maxSimpleNo;
     }
 
@@ -60,19 +60,19 @@ public class SuperDto {
         this.max = max;
     }
 
-    public byte getMinSystemNo() {
+    public int getMinSystemNo() {
         return minSystemNo;
     }
 
-    public void setMinSystemNo(byte minSystemNo) {
+    public void setMinSystemNo(int minSystemNo) {
         this.minSystemNo = minSystemNo;
     }
 
-    public byte getMinSimpleNo() {
+    public int getMinSimpleNo() {
         return minSimpleNo;
     }
 
-    public void setMinSimpleNo(byte minSimpleNo) {
+    public void setMinSimpleNo(int minSimpleNo) {
         this.minSimpleNo = minSimpleNo;
     }
 
@@ -84,51 +84,51 @@ public class SuperDto {
         this.min = min;
     }
 
-    public byte getMaxTemperatureSystemNo() {
+    public int getMaxTemperatureSystemNo() {
         return maxTemperatureSystemNo;
     }
 
-    public void setMaxTemperatureSystemNo(byte maxTemperatureSystemNo) {
+    public void setMaxTemperatureSystemNo(int maxTemperatureSystemNo) {
         this.maxTemperatureSystemNo = maxTemperatureSystemNo;
     }
 
-    public byte getMaxSeq() {
+    public int getMaxSeq() {
         return maxSeq;
     }
 
-    public void setMaxSeq(byte maxSeq) {
+    public void setMaxSeq(int maxSeq) {
         this.maxSeq = maxSeq;
     }
 
-    public byte getMaxTemperature() {
+    public int getMaxTemperature() {
         return maxTemperature;
     }
 
-    public void setMaxTemperature(byte maxTemperature) {
+    public void setMaxTemperature(int maxTemperature) {
         this.maxTemperature = maxTemperature;
     }
 
-    public byte getMinTemperatureSystemNo() {
+    public int getMinTemperatureSystemNo() {
         return minTemperatureSystemNo;
     }
 
-    public void setMinTemperatureSystemNo(byte minTemperatureSystemNo) {
+    public void setMinTemperatureSystemNo(int minTemperatureSystemNo) {
         this.minTemperatureSystemNo = minTemperatureSystemNo;
     }
 
-    public byte getMinSeq() {
+    public int getMinSeq() {
         return minSeq;
     }
 
-    public void setMinSeq(byte minSeq) {
+    public void setMinSeq(int minSeq) {
         this.minSeq = minSeq;
     }
 
-    public byte getMinTemperature() {
+    public int getMinTemperature() {
         return minTemperature;
     }
 
-    public void setMinTemperature(byte minTemperature) {
+    public void setMinTemperature(int minTemperature) {
         this.minTemperature = minTemperature;
     }
 }
