+++ jhpt-server/src/main/java/com/tyj/jhpt/server/body/dto/WenDuDto.java
@@ -12,17 +12,17 @@ package com.tyj.jhpt.server.body.dto;
  */
 public class WenDuDto {
     // 可充电储能子系统号
-    private byte systemNo;
+    private int systemNo;
     // 可充电储能温度探针个数
     private int count;
     // 可充电储能子系统各温度探针监测到的温度值
     private byte[] values;
 
-    public byte getSystemNo() {
+    public int getSystemNo() {
         return systemNo;
     }
 
-    public void setSystemNo(byte systemNo) {
+    public void setSystemNo(int systemNo) {
         this.systemNo = systemNo;
     }
